% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{bronchitis}
\alias{bronchitis}
\title{Bronchitis}
\format{
\describe{
\item{bron}{Presence of bronchitis (0 for no and 1 for yes)}
\item{cigs}{Average daily number of smoked cigarettes}
\item{poll}{Pollution index}

}
}
\usage{
bronchitis
}
\description{
Data collected in a study to assess the effects of smoking and pollution on being diagnosed with bronchitis. This dataset is based on 212 subjects.
}
\keyword{datasets}
