% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{cortisol}
\alias{cortisol}
\title{biomarkers in pigs fed with various diets}
\format{
A data frame with 61 rows and 9 variables:
\describe{
  \item{id}{the id of the pig}
  \item{group}{the diet fed to the pig (chipped diet or non chipped diet)}
  \item{gender}{the gender of the pig}
  \item{urine_cortisol_pg_mg}{urine costisol in pg/ml}
  \item{serum_cortisol_ng_ml}{serum cortisol in ng/ml}
  \item{serum_acth_pg_ml}{serum acth in pg/ml}
  \item{serum_crh_pg_ml}{serum crh in pg/ml}
  \item{testosterone_ng_ml}{testosterone in ng/ml}
  \item{lh_ng_ml}{LH in ng/ml}
  }
}
\usage{
cortisol
}
\description{
A dataset containing measured biomarkers in pigs fed with various diets
}
\keyword{datasets}
