% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{diet}
\alias{diet}
\title{Diet}
\format{
\describe{
  \item{id}{ID}
  \item{gender}{Gender (male or female)}
  \item{age}{Age in years}
  \item{height}{Height in m}
  \item{diet.type}{Type of diet (A, B or C)}
  \item{initial.weight}{Initial weight in kg}
  \item{final.weight}{Final weight in kg}
}
}
\usage{
diet
}
\description{
Diet
}
\keyword{datasets}
