% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{fev}
\alias{fev}
\title{fev}
\format{
\describe{
\item{fev}{forced expiratory volume or FEV, which measures the amount of air a person can exhale during a forced breath.}
\item{age}{age in years}
\item{sex}{gender of the person (0 for males and 1 for females).}
\item{height}{height in cm}
\item{smoke}{smoking behavior (0 for non-smokers and 1 for smokers).}
}
}
\usage{
fev
}
\description{
Study conducted in suburban Boston in the late 1970s to investigate the relationship between forced expiratory volume and smoking behavior in 654 youths between the ages of 3 and 19.

fev
}
\keyword{datasets}
