% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{HP13Cbicarbonate}
\alias{HP13Cbicarbonate}
\title{HP13Cbicarbonate}
\format{
\describe{
  \item{signal}{HP13C bicarbonate signal intensities normalized to the total sum of metabolites}
  \item{dose}{initial reaction rate as a function of the injected dose of HP13C pyruvate}
  \item{group}{fed and overnight-fasted}
}
}
\source{
\url{https://www.nature.com/articles/s42003-021-02978-2}
}
\usage{
HP13Cbicarbonate
}
\description{
Data from an experiment made on rats which compares the HP13C bicarbonate signal intensities normalized to
the total sum of metabolites and corresponding initial reaction rate as a function of the injected dose of HP1-13C pyruvate.
Two groups of rats were compared (i.e. fed and overnight-fasted). Dataset from Can et al. 2022.
}
\keyword{datasets}
