% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{kuwait_bp}
\alias{kuwait_bp}
\title{Kuwait Blood Pressure}
\format{
\describe{
\item{age}{Age in years}
\item{weight}{Weight in kg}
\item{height}{Height in mm}
\item{chin}{Chin skinfold in cm}
\item{forearm}{Forearm skinfold in cm}
\item{calf}{Calf skinfold in cm}
\item{pulse}{Resting pulse rate}
\item{left_handed}{Whether or not the participant is left-handed}
\item{bmi}{The Body Mass Index (BMI) of the participant}
\item{systol}{Systolic blood pressure}
}
}
\usage{
kuwait_bp
}
\description{
This dataset contains a collection of variables believed to be potentially associated with the blood pressure measurements of 213 individuals from Kuwait. The dataset lists the following variables:
}
\keyword{datasets}
