% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{PeruvianBP}
\alias{PeruvianBP}
\title{Peruvian Blood Pressure}
\format{
\describe{
\item{Age}{Age in years}
\item{Years}{Years in urban area}
\item{Weight}{Weight in kg}
\item{Height}{Height in mm}
\item{Chin}{Chin skinfold}
\item{Forearm}{Forearm skinfold}
\item{Calf}{Calf skinfold}
\item{Pulse}{Resting pulse rate}
\item{Systol}{Systolic blood pressure}
}
}
\usage{
PeruvianBP
}
\description{
This dataset consists of variables possibly relating to blood pressures of 39 Peruvians who have moved from rural high-altitude areas to urban lower-altitude areas.
}
\keyword{datasets}
