% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{pm_exposure}
\alias{pm_exposure}
\title{Particulate Matter Exposure and Lung Function}
\format{
A data frame with 200 rows and 7 variables:
\describe{
\item{Age}{Age of the participant in years}
\item{Height}{Height of the participant in cm}
\item{Weight}{Weight of the participant in kg}
\item{PM2_5_Exposure}{Estimated 5-years average exposure to fine particulate matter (PM2.5), measured in micrograms per cubic meter}
\item{Rural}{A binary indicator of residential location (1 = Rural, 0 = Non-rural)}
\item{AreaType}{A factor indicating the specific type of residential area ("rural", "suburban", or "urban")}
\item{FEV1}{Forced Expiratory Volume in 1 second (measured in Liters), representing lung function}
}
}
\usage{
pm_exposure
}
\description{
This dataset contains demographic, environmental and respiratory data for 200 male adults (ages 40–75) to analyze factors affecting lung capacity. It contrasts populations across three distinct air-quality environments: a pristine high-altitude rural town, a suburban area with moderate air quality, and a highly industrialized urban region.
}
\keyword{datasets}
