% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{reading}
\alias{reading}
\title{reading}
\format{
\describe{
  \item{id}{id}
  \item{score}{score}
  \item{age}{age}
  \item{group}{group}
}
}
\usage{
reading
}
\description{
reading
}
\keyword{datasets}
