% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{snoring}
\alias{snoring}
\title{snoring}
\format{
\describe{
\item{sex}{gender of the person (0 for males and 1 for females).}
\item{age}{age in years}
\item{height}{height in cm}
\item{weight}{weight in kg}
\item{smoke}{smoking behavior (0 for non-smokers and 1 for smokers).}
\item{alcohol}{number of glasses drunk per day (in red wine equivalent).}
\item{snore}{snoring diagnosis (0=not snoring, 1=snoring).}
}
}
\usage{
snoring
}
\description{
Study on the physical and behavioral characteristics of snorers.

snoring
}
\keyword{datasets}
