% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_w_points.R
\name{boxplot_w_points}
\alias{boxplot_w_points}
\title{boxplot_w_points}
\usage{
boxplot_w_points(
  ...,
  col = "#FFB3DACC",
  col_boxplot = "#d2d2d2",
  horizontal = F,
  main = "",
  names,
  log = "",
  las = 0
)
}
\arguments{
\item{col}{color of the points to be added to the boxplot}

\item{horizontal}{logical indicating if the boxplots should be horizontal; default FALSE means vertical boxes.}

\item{main}{string indicating the title of the plot}

\item{data}{a vector of numeric values}
}
\description{
boxplot_w_points
}
\examples{
\dontrun{
x = rnorm(20, mean=5)
y = rnorm(20, mean=10)
z = rnorm(20, mean=15)
boxplot_w_points(x, main = "test")
boxplot_w_points(x,y,z, names = c("x","y", "z"), horizontal = F, las=1, main="data")
}
}
