% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_w_points.R
\name{boxplot_w_points}
\alias{boxplot_w_points}
\title{boxplot_w_points}
\usage{
boxplot_w_points(data, col = "#CCCCCC7F", horiz = F, main = "", nticks = 8L)
}
\arguments{
\item{data}{a vector of numeric values}

\item{col}{color of the points to be added to the boxplot}

\item{horiz}{logical indicating if the boxplots should be horizontal; default FALSE means vertical boxes.}

\item{main}{string indicating the title of the plot}

\item{nticks}{an integer specifying the number of ticks to show}
}
\description{
boxplot_w_points
}
\examples{
\dontrun{
set.seed(12456)
boxplot_w_points(data = rnorm(100), horiz = TRUE, main="data")
set.seed(12456)
boxplot_w_points(data = rnorm(100), horiz = FALSE, main="data")
}
}
