% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{BreastCancer}
\alias{BreastCancer}
\title{Breast Cancer}
\format{
\describe{
\item{Age}{Age in years}
\item{BMI}{Body mass index in kg/\eqn{m^2}}
\item{Glucose}{Glucose in mg/dL}
\item{Insulin}{Insulin in \eqn{\mu}U/mL}
\item{HOMA}{Homeostasis model assessment}
\item{Classification}{Presence of breast cancer (0 if no cancer, 1 if with cancer)}
}
}
\source{
\url{https://bmccancer.biomedcentral.com/articles/10.1186/s12885-017-3877-1}
}
\usage{
BreastCancer
}
\description{
This dataset consists of several clinical features observed or measured for 116 participants in a study of breast cancer.
}
\references{
Patricio, Miguel, et al. "Using Resistin, glucose, age and BMI to predict the presence of breast cancer", BMC Cancer, (2018).
}
\keyword{datasets}
