% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{codex}
\alias{codex}
\title{codex}
\format{
\describe{
  \item{id}{id}
  \item{gender}{score}
  \item{age}{age}
  \item{bmi}{height}
  \item{weight}{diet.type}
  \item{number_doses}{initial.weight}
  \item{tmax}{final.weight}
  \item{cmax}{id}
  \item{t1_2}{score}
  \item{auc}{age}
  \item{lenght_hospital}{height}
  \item{lenght_intermed}{diet.type}
  \item{crp}{initial.weight}
  \item{comor_e}{final.weight}
  \item{comor_p}{final.weight}
  \item{comor_v}{final.weight}
  \item{comor_c}{final.weight}
  \item{comor_r}{final.weight}
  \item{obese}{final.weight}
}
}
\usage{
codex
}
\description{
codex
}
\keyword{datasets}
