% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{cortisol}
\alias{cortisol}
\title{biomarkers in pigs fed with various diets}
\format{
A data frame with 61 rows and 9 variables:
\describe{
  \item{id}{the id of the pig}
  \item{group}{the diet fed to the pig (chipped diet or non chipped diet)}
  \item{gender}{the gender of the pig}
  \item{cortisol}{urine costisol in pg/ml}
  \item{acth}{serum acth in pg/ml}
  \item{crh}{serum crh in pg/ml}
  \item{testosterone}{testosterone in ng/ml}
  \item{lh}{LH in ng/ml}
  \item{caloric}{Daily caloric intake in kcal}
  }
}
\usage{
cortisol
}
\description{
A dataset containing measured biomarkers in pigs fed with various diets
}
\keyword{datasets}
