% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{diet}
\alias{diet}
\title{diet}
\format{
\describe{
  \item{id}{id}
  \item{gender}{score}
  \item{age}{age}
  \item{height}{height}
  \item{diet.type}{diet.type}
  \item{initial.weight}{initial.weight}
  \item{final.weight}{final.weight}
}
}
\usage{
diet
}
\description{
diet
}
\keyword{datasets}
