% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_to_normal.R
\name{hist_compare_to_normal}
\alias{hist_compare_to_normal}
\title{hist_compare_to_normal}
\usage{
hist_compare_to_normal(
  x,
  col = "lightgray",
  main = "",
  xlab = "",
  ylab = "",
  lwd_line = 1.5,
  col_line = "#ff160e",
  add_estimated_param = T,
  legend_position = "topleft",
  ...
)
}
\arguments{
\item{x}{data vector to be visualized}

\item{col}{color of the histogram}

\item{main}{string indicating the title of the plot}

\item{xlab}{a string indicating the x label}

\item{ylab}{a string indicating the y label}

\item{add_estimated_param}{a Boolean if the estimated parameters of the Normal distribution should be plotted}

\item{legend_position}{a string specifying the position of the legend.}

\item{las}{a numeric value indicating the orientation of the tick mark labels and any other text added to a plot after its initialization. The options are as follows: always parallel to the axis (the default, 0), always horizontal (1), always perpendicular to the axis (2), and always vertical (3).}
}
\description{
hist_compare_to_normal
}
\examples{
n <- 1000
x <- rnorm(n = n)
hist_compare_to_normal(x)
x2 <- rexp(n, rate = 25)
hist_compare_to_normal(x2, legend_position = "topright")
}
