% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{pharmacy}
\alias{pharmacy}
\title{Customer attendance of a pharmacy in Geneva}
\format{
A data frame with 17520 rows and 4 variables:
\describe{
  \item{date}{the date}
  \item{hours}{the hour of the day}
  \item{weekday}{the week day}
  \item{attendance}{the recorded number of clients}
}
}
\usage{
pharmacy
}
\description{
This dataset contains the number of clients in a pharmacy for each hour over two years.
}
\keyword{datasets}
