% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{reading}
\alias{reading}
\title{Reading dataset}
\format{
\describe{
  \item{id}{Student id.}
  \item{score}{Degree of Reading Power (DRP) test score.}
  \item{age}{Age of the students.}
  \item{group}{Binary variable indicating whether a student participated to the directed reading activities (Treatment if the student participated, Control otherwise).}
}
}
\usage{
reading
}
\description{
Study on the effectiveness of directed reading activities for elementary school students (6-12 years old).
}
\keyword{datasets}
